adds images (but they are visible?)

+++ build.py
@@ -1,10 +1,17 @@
 import os
 import re
 import json
+import shutil
 
 AUTHORS = ["anonymous"]
 
+status = os.system("rm -rf ./build")
+if status != 0:
+    print("[ERROR]: Failed to clean build dir")
+    exit(1)
+
 os.mkdir("./build") if not os.path.exists("./build") else None
+os.mkdir("./build/img") if not os.path.exists("./build/img") else None
 
 status = os.system("rm -rf sprig && git clone https://github.com/hackclub/sprig --depth 1 --branch main")
 if status != 0:
@@ -12,6 +19,9 @@ if status != 0:
     exit(1)
 
 metadata = {}
+
+shutil.copytree("./css", "./build/css") 
+shutil.copytree("./scripts", "./build/scripts")
 
 for i in os.listdir("./sprig/games"):
     if os.path.isdir(f"./sprig/games/{i}"):
@@ -40,8 +50,22 @@ for i in os.listdir("./sprig/games"):
             os.mkdir("./build/games") if not os.path.exists("./build/games") else None
             with open(f"./build/games/{i}", "w", encoding="utf-8") as outFile:
                 outFile.write(content)
+            
+            game_file_basename = i.replace(".js","")
+            img_file_path = f"./sprig/games/img/{game_file_basename}.png"
+            dest = f"./build/img/{game_file_basename}.png"
+            
+            has_img = False
+            if os.path.exists(img_file_path):
+                print("copyfile", img_file_path, dest)
+                shutil.copyfile(img_file_path, dest)
+                has_img = True
+            
+
             metadata[i] = {
                 "title": title,
+                "game_file_basename": game_file_basename,
+                "has_img": has_img,
                 "author": author,
                 "addedOn": addedOn,
                 "description": description
